fix: Remove manual setting of URL in sigplot_ext.js (#24)
Manually setting the URL referencing '/files/...'
to plot requires too much complex logic on the
server side to keep track of where the server root directory
is located vs where the notebook is running.

Instead, simply pass the name of the file relative to the
notebook, and the client should figure it out when it sends
the request.